# docs/api/openapi/trip-service.openapi.yaml
openapi: 3.0.3
info:
  title: Travel AI Trip Service API
  version: 1.0.0
  description: Core frontend-facing trip contract. It models existing raw response bodies while new endpoints move toward the documented envelope.
servers:
  - url: /api/trips
security: [{ BearerAuth: [] }]
...
paths:
...
  /trips/{tripId}/invitations:
    get:
      tags: [Collaboration]
      operationId: listTripInvitations
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      responses:
        '200':
          description: Owner-visible invitation list.
          content: { application/json: { schema: { $ref: '#/components/schemas/ListTripInvitationsResponse' } } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        default: { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Collaboration]
      operationId: createTripInvitation
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateTripInvitationRequest' } } }
      responses:
        '201':
          description: Created invitation.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripInvitation' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/invitations/{invitationId}/accept:
    post:
      tags: [Collaboration]
      operationId: acceptTripInvitation
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Accepted invitation.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripInvitation' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/invitations/{invitationId}/decline:
    post:
      tags: [Collaboration]
      operationId: declineTripInvitation
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Declined invitation.
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/invitations/{invitationId}/resend:
    post:
      tags: [Collaboration]
      operationId: resendTripInvitation
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/ResendTripInvitationRequest' } } }
      responses:
        '200':
          description: Resent invitation.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripInvitation' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/invitations/{invitationId}/revoke:
    post:
      tags: [Collaboration]
      operationId: revokeTripInvitation
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Revoked invitation.
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/members:
    get:
      tags: [Collaboration]
      operationId: listTripMembers
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      responses:
        '200':
          description: Trip member roster.
          content: { application/json: { schema: { $ref: '#/components/schemas/ListTripMembersResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/members/transfer-ownership:
    post:
      tags: [Collaboration]
      operationId: transferTripOwnership
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [newOwnerUserId], properties: { newOwnerUserId: { type: string, format: uuid } } } } }
      responses:
        '200': { $ref: '#/components/responses/TripResponse' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/members/leave:
    post:
      tags: [Collaboration]
      operationId: leaveTrip
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      responses:
        '200':
          description: Left trip.
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /trips/{tripId}/suggestions:
    get:
      tags: [Collaboration]
      operationId: listTripSuggestions
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: status, in: query, schema: { $ref: '#/components/schemas/TripSuggestionStatus' } }]
      responses:
        '200':
          description: Trip suggestions.
          content: { application/json: { schema: { $ref: '#/components/schemas/ListTripSuggestionsResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Collaboration]
      operationId: createTripSuggestion
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateTripSuggestionRequest' } } }
      responses:
        '201':
          description: Created suggestion.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripSuggestion' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/suggestions/{suggestionId}/accept:
    post:
      tags: [Collaboration]
      operationId: acceptTripSuggestion
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: suggestionId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { expectedItineraryRevision: { type: integer } } } } }
      responses:
        '200':
          description: Accepted suggestion.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripSuggestion' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/suggestions/{suggestionId}/reject:
    post:
      tags: [Collaboration]
      operationId: rejectTripSuggestion
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: suggestionId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Rejected suggestion.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripSuggestion' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/suggestions/{suggestionId}/resolve:
    post:
      tags: [Collaboration]
      operationId: resolveTripSuggestion
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: suggestionId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Resolved suggestion.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripSuggestion' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/votes:
    get:
      tags: [Collaboration]
      operationId: listTripVotes
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: targetType, in: query, schema: { $ref: '#/components/schemas/TripVoteTargetType' } }, { name: targetId, in: query, schema: { type: string } }]
      responses:
        '200':
          description: Vote summaries.
          content: { application/json: { schema: { $ref: '#/components/schemas/ListTripVoteSummariesResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Collaboration]
      operationId: setTripVote
      parameters: [{ $ref: '#/components/parameters/TripId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SetTripVoteRequest' } } }
      responses:
        '200':
          description: Updated vote summary.
          content: { application/json: { schema: { $ref: '#/components/schemas/TripVoteSummary' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /trips/{tripId}/votes/{voteId}:
    delete:
      tags: [Collaboration]
      operationId: deleteTripVote
      parameters: [{ $ref: '#/components/parameters/TripId' }, { name: voteId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Deleted vote.
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TripId: { name: tripId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    TripAccess:
      type: object
      required: [role, canEdit, canManageCollaborators, canManageShare, canRestoreVersion, canDelete]
      properties:
        role: { type: string, enum: [owner, editor, viewer] }
        source: { type: string, enum: [owner, workspace, collaborator, public] }
        canEdit: { type: boolean }
        canManageCollaborators: { type: boolean }
        canManageShare: { type: boolean }
        canRestoreVersion: { type: boolean }
        canDelete: { type: boolean }
    Trip:
      type: object
      required: [id, destination, days, budgetCurrency, travelers, interests, pace, status, itineraryRevision, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid, nullable: true }
        workspaceId: { type: string, format: uuid, nullable: true }
        scope: { type: string, enum: [personal, workspace] }
        tripType: { type: string, enum: [single_destination, multi_destination] }
        route: { type: object, nullable: true, additionalProperties: true }
        destination: { type: string }
        startDate: { type: string, format: date, nullable: true }
        days: { type: integer, minimum: 1 }
        budgetAmount: { type: number, nullable: true }
        budgetCurrency: { type: string, pattern: '^[A-Z]{3}$' }
        budget: { type: object, nullable: true, additionalProperties: true }
        accommodation: { type: object, nullable: true, additionalProperties: true }
        travelers: { type: integer, minimum: 1 }
        interests: { type: array, items: { type: string } }
        pace: { type: string }
        status: { type: string, enum: [DRAFT, PROCESSING, COMPLETED, FAILED] }
        itinerary: { type: object, nullable: true, additionalProperties: true }
        itineraryRevision: { type: integer, minimum: 0 }
        lifecycle: { type: string, enum: [draft, planning, ready, active, completed, archived] }
        archivedAt: { type: string, format: date-time, nullable: true }
        access:
          allOf: [{ $ref: '#/components/schemas/TripAccess' }]
          nullable: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CollaboratorRole:
      type: string
      enum: [viewer, editor]
    TripInvitationStatus:
      type: string
      enum: [pending, accepted, declined, expired, revoked]
    CreateTripInvitationRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/CollaboratorRole' }
        message: { type: string, maxLength: 500 }
        expiresAt: { type: string, format: date-time, nullable: true }
    ResendTripInvitationRequest:
      type: object
      properties:
        message: { type: string, maxLength: 500 }
        expiresAt: { type: string, format: date-time, nullable: true }
    TripInvitation:
      type: object
      required: [id, tripId, inviterUserId, email, role, status, expiresAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        tripId: { type: string, format: uuid }
        inviterUserId: { type: string, format: uuid }
        invitedUserId: { type: string, format: uuid, nullable: true }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/CollaboratorRole' }
        status: { $ref: '#/components/schemas/TripInvitationStatus' }
        message: { type: string }
        expiresAt: { type: string, format: date-time }
        acceptedAt: { type: string, format: date-time, nullable: true }
        declinedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ListTripInvitationsResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TripInvitation' } }
    TripMember:
      type: object
      required: [userId, role, status, permissions, isSelf]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email, nullable: true }
        displayName: { type: string, nullable: true }
        role: { type: string }
        status: { type: string, enum: [active, invited, removed] }
        joinedAt: { type: string, format: date-time, nullable: true }
        invitedBy: { type: string, format: uuid, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        permissions: { type: object, additionalProperties: { type: boolean } }
        isSelf: { type: boolean }
    ListTripMembersResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TripMember' } }
...
    TripSuggestionType:
      type: string
      enum: [activity_replacement, time_change, budget_adjustment, route_change, note]
    TripSuggestionTargetType:
      type: string
      enum: [trip, day, itinerary_item, budget_item, route, attachment]
    TripSuggestionStatus:
      type: string
      enum: [open, accepted, rejected, resolved]
    CreateTripSuggestionRequest:
      type: object
      required: [suggestionType, targetType]
      properties:
        suggestionType: { $ref: '#/components/schemas/TripSuggestionType' }
        targetType: { $ref: '#/components/schemas/TripSuggestionTargetType' }
        targetId: { type: string, maxLength: 200 }
        before: {}
        after: {}
        comment: { type: string, maxLength: 2000 }
        metadata: { type: object, additionalProperties: true }
    TripSuggestion:
      type: object
      required: [id, tripId, authorUserId, suggestionType, targetType, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        tripId: { type: string, format: uuid }
        authorUserId: { type: string, format: uuid }
        suggestionType: { $ref: '#/components/schemas/TripSuggestionType' }
        targetType: { $ref: '#/components/schemas/TripSuggestionTargetType' }
        targetId: { type: string }
        status: { $ref: '#/components/schemas/TripSuggestionStatus' }
        before: {}
        after: {}
        comment: { type: string }
        metadata: { type: object, additionalProperties: true }
        appliedItineraryRevision: { type: integer, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        resolvedByUserId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isAuthor: { type: boolean }
        canResolve: { type: boolean }
    ListTripSuggestionsResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TripSuggestion' } }
    TripVoteTargetType:
      type: string
      enum: [activity, restaurant, hotel, destination, suggestion]
    TripVoteType:
      type: string
      enum: [thumbs_up, thumbs_down, heart, star]
    SetTripVoteRequest:
      type: object
      required: [targetType, targetId, voteType]
      properties:
        targetType: { $ref: '#/components/schemas/TripVoteTargetType' }
        targetId: { type: string, minLength: 1, maxLength: 200 }
        voteType: { $ref: '#/components/schemas/TripVoteType' }
        metadata: { type: object, additionalProperties: true }
    TripVoteSummary:
      type: object
      required: [targetType, targetId, counts]
      properties:
        targetType: { $ref: '#/components/schemas/TripVoteTargetType' }
        targetId: { type: string }
        counts: { type: object, additionalProperties: { type: integer } }
        currentVote: { $ref: '#/components/schemas/TripVoteType' }
    ListTripVoteSummariesResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/TripVoteSummary' } }
    SuccessResponse:
      type: object
      required: [success]
      properties: { success: { type: boolean } }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              required: [code, message]
              properties:
                code: { type: string }
                message: { type: string }
                details:
                  oneOf:
                    - { type: array, items: { type: object, additionalProperties: true } }
                    - { type: object, additionalProperties: true }
                requestId: { type: string }
        currentItineraryRevision: { type: integer }
  responses:
    TripResponse:
      description: Private trip DTO.
      content: { application/json: { schema: { $ref: '#/components/schemas/Trip' } } }
...
    ErrorResponse:
      description: User-safe request failure.
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }